+++ front/src/pages/Friend/PendingFriend.tsx
@@ -4,6 +4,7 @@ import Friend from './interface/friendDto';
 import { useFriends } from './useFriends';
 import FriendCard from "./FriendCard";
 import { FriendStatus } from './interface/friendStatus';
+import PendingFriendCard from './PendingFriendCard';
 
 const PendingFriend = () => {
 	const [friends, setFriends] = useState<Friend[]>([]);
@@ -33,7 +34,7 @@ const PendingFriend = () => {
 
 				<div className="h-full m-2.5 bg-black/10 rounded-md	shadow-md shadow-white box-border justify-center items-center overflow-y-auto max-h-[80%]">
 					{friends?.map((friend: Friend) => (
-						<FriendCard key={friend.id} friend={friend} />
+						<PendingFriendCard key={friend.id} friend={friend} />
 					))}
 				</div>
 			)
