+++ front/src/pages/Friend/interface/friendDto.tsx
