+++ front/src/pages/Friend/interface/friendStatus.tsx
