+++ front/src/pages/Friend/status.tsx
